requirements.txt
python-telegram-bot==20.7
MetaTrader5==5.0.45
pandas==2.1.4
numpy==1.24.4
scikit-learn
ta==0.10.2
python-dotenv==1.0.0
schedule
psutil==5.9.6
requests
hupper==1.12.1
watchdog==6.0.0
